Clean up, add comments, merge migrations
