--- alembic/versions/60adeee6f30b_job_series_view_clusters_join_left_.py
@@ -1,44 +1,0 @@
-"""job_series_view: clusters join LEFT (droppable when cluster_name unused)
-
-Revision ID: 60adeee6f30b
-Revises: d64ccf9cd902
-Create Date: 2026-07-17 01:34:35.760653+00:00
-
-"""
-
-from typing import Sequence, Union
-
-from alembic_utils.pg_view import PGView
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision: str = "60adeee6f30b"
-down_revision: Union[str, Sequence[str], None] = "d64ccf9cd902"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    public_job_series_view = PGView(
-        schema="public",
-        signature="job_series_view",
-        definition="SELECT slurm_jobs.id AS job_db_id, slurm_jobs.sarc_user_id AS sarc_user_id, slurm_jobs.cluster_id, slurm_jobs.account, slurm_jobs.job_id, slurm_jobs.array_job_id, slurm_jobs.task_id, slurm_jobs.name, slurm_jobs.cluster_user, slurm_jobs.\"group\", slurm_jobs.job_state, slurm_jobs.exit_code, slurm_jobs.signal, slurm_jobs.partition, slurm_jobs.nodes, slurm_jobs.work_dir, slurm_jobs.submit_line, slurm_jobs.constraints, slurm_jobs.priority, slurm_jobs.qos, slurm_jobs.\"CLEAR_SCHEDULING\", slurm_jobs.\"STARTED_ON_SUBMIT\", slurm_jobs.\"STARTED_ON_SCHEDULE\", slurm_jobs.\"STARTED_ON_BACKFILL\", slurm_jobs.time_limit, slurm_jobs.submit_time, slurm_jobs.start_time, slurm_jobs.end_time, slurm_jobs.elapsed_time, slurm_jobs.requested_cpu, slurm_jobs.requested_mem, slurm_jobs.requested_node, slurm_jobs.requested_billing, slurm_jobs.requested_gres_gpu, slurm_jobs.requested_gpu_type, slurm_jobs.allocated_cpu, slurm_jobs.allocated_mem, slurm_jobs.allocated_node, slurm_jobs.allocated_billing, slurm_jobs.allocated_gres_gpu, slurm_jobs.allocated_gpu_type, slurm_jobs.harmonized_gpu_type, users.display_name, users.email, clusters.name AS cluster_name, (SELECT membertypedb.member_type \nFROM membertypedb \nWHERE membertypedb.user_id = slurm_jobs.sarc_user_id AND membertypedb.valid @> slurm_jobs.submit_time) AS member_type, (SELECT json_object_agg(jobstatisticdb.name, json_build_object('mean', jobstatisticdb.mean, 'std', jobstatisticdb.std, 'q05', jobstatisticdb.q05, 'q25', jobstatisticdb.q25, 'median', jobstatisticdb.median, 'q75', jobstatisticdb.q75, 'max', jobstatisticdb.max, 'unused', jobstatisticdb.unused)) AS json_object_agg_1 \nFROM jobstatisticdb \nWHERE jobstatisticdb.job_id = slurm_jobs.id) AS statistics, (SELECT json_agg(supervisorshelper.supervisor ORDER BY supervisorshelper.pos) AS json_agg_1 \nFROM user_supervisors JOIN supervisorshelper ON user_supervisors.id = supervisorshelper.list_id \nWHERE user_supervisors.user_id = slurm_jobs.sarc_user_id AND user_supervisors.valid @> slurm_jobs.submit_time) AS supervisors, gpurgudb.rgu AS gpu_type_rgu, gpurgudb.drac_rgu AS gpu_type_rgu_drac, coalesce(slurm_jobs.requested_gres_gpu, 0) * gpurgudb.rgu AS requested_rgu, coalesce(slurm_jobs.requested_gres_gpu, 0) * gpurgudb.drac_rgu AS requested_rgu_drac, coalesce(slurm_jobs.allocated_gres_gpu, 0) * gpurgudb.rgu AS allocated_rgu, coalesce(slurm_jobs.allocated_gres_gpu, 0) * gpurgudb.drac_rgu AS allocated_rgu_drac, slurm_jobs.elapsed_time * slurm_jobs.requested_cpu AS requested_cpu_cost, (1 - jobstatisticdb_1.mean) * slurm_jobs.elapsed_time * slurm_jobs.requested_cpu AS requested_cpu_waste, slurm_jobs.elapsed_time * slurm_jobs.allocated_cpu AS allocated_cpu_cost, (1 - jobstatisticdb_1.mean) * slurm_jobs.elapsed_time * slurm_jobs.allocated_cpu AS allocated_cpu_waste, slurm_jobs.elapsed_time * (slurm_jobs.allocated_cpu - slurm_jobs.requested_cpu) AS cpu_overbilling_cost, slurm_jobs.elapsed_time * slurm_jobs.requested_gres_gpu * gpurgudb.drac_rgu AS requested_gpu_cost, (1 - jobstatisticdb_2.mean) * slurm_jobs.elapsed_time * slurm_jobs.requested_gres_gpu * gpurgudb.drac_rgu AS requested_gpu_waste, slurm_jobs.elapsed_time * slurm_jobs.allocated_gres_gpu * gpurgudb.drac_rgu AS allocated_gpu_cost, (1 - jobstatisticdb_2.mean) * slurm_jobs.elapsed_time * slurm_jobs.allocated_gres_gpu * gpurgudb.drac_rgu AS allocated_gpu_waste, slurm_jobs.elapsed_time * (slurm_jobs.allocated_gres_gpu - slurm_jobs.requested_gres_gpu) * gpurgudb.drac_rgu AS gpu_overbilling_cost \nFROM slurm_jobs LEFT OUTER JOIN users ON slurm_jobs.sarc_user_id = users.id LEFT OUTER JOIN clusters ON slurm_jobs.cluster_id = clusters.id LEFT OUTER JOIN gpurgudb ON gpurgudb.name = slurm_jobs.harmonized_gpu_type LEFT OUTER JOIN jobstatisticdb AS jobstatisticdb_2 ON jobstatisticdb_2.job_id = slurm_jobs.id AND jobstatisticdb_2.name = 'gpu_sm_occupancy' LEFT OUTER JOIN jobstatisticdb AS jobstatisticdb_1 ON jobstatisticdb_1.job_id = slurm_jobs.id AND jobstatisticdb_1.name = 'cpu_utilization'",
-    )
-    op.replace_entity(public_job_series_view)
-
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    public_job_series_view = PGView(
-        schema="public",
-        signature="job_series_view",
-        definition="SELECT slurm_jobs.id AS job_db_id,\n    slurm_jobs.sarc_user_id,\n    slurm_jobs.cluster_id,\n    slurm_jobs.account,\n    slurm_jobs.job_id,\n    slurm_jobs.array_job_id,\n    slurm_jobs.task_id,\n    slurm_jobs.name,\n    slurm_jobs.cluster_user,\n    slurm_jobs.\"group\",\n    slurm_jobs.job_state,\n    slurm_jobs.exit_code,\n    slurm_jobs.signal,\n    slurm_jobs.partition,\n    slurm_jobs.nodes,\n    slurm_jobs.work_dir,\n    slurm_jobs.submit_line,\n    slurm_jobs.constraints,\n    slurm_jobs.priority,\n    slurm_jobs.qos,\n    slurm_jobs.\"CLEAR_SCHEDULING\",\n    slurm_jobs.\"STARTED_ON_SUBMIT\",\n    slurm_jobs.\"STARTED_ON_SCHEDULE\",\n    slurm_jobs.\"STARTED_ON_BACKFILL\",\n    slurm_jobs.time_limit,\n    slurm_jobs.submit_time,\n    slurm_jobs.start_time,\n    slurm_jobs.end_time,\n    slurm_jobs.elapsed_time,\n    slurm_jobs.requested_cpu,\n    slurm_jobs.requested_mem,\n    slurm_jobs.requested_node,\n    slurm_jobs.requested_billing,\n    slurm_jobs.requested_gres_gpu,\n    slurm_jobs.requested_gpu_type,\n    slurm_jobs.allocated_cpu,\n    slurm_jobs.allocated_mem,\n    slurm_jobs.allocated_node,\n    slurm_jobs.allocated_billing,\n    slurm_jobs.allocated_gres_gpu,\n    slurm_jobs.allocated_gpu_type,\n    slurm_jobs.harmonized_gpu_type,\n    users.display_name,\n    users.email,\n    clusters.name AS cluster_name,\n    ( SELECT membertypedb.member_type\n           FROM membertypedb\n          WHERE ((membertypedb.user_id = slurm_jobs.sarc_user_id) AND (membertypedb.valid @> slurm_jobs.submit_time))) AS member_type,\n    ( SELECT json_object_agg(jobstatisticdb.name, json_build_object('mean', jobstatisticdb.mean, 'std', jobstatisticdb.std, 'q05', jobstatisticdb.q05, 'q25', jobstatisticdb.q25, 'median', jobstatisticdb.median, 'q75', jobstatisticdb.q75, 'max', jobstatisticdb.max, 'unused', jobstatisticdb.unused)) AS json_object_agg_1\n           FROM jobstatisticdb\n          WHERE (jobstatisticdb.job_id = slurm_jobs.id)) AS statistics,\n    ( SELECT json_agg(supervisorshelper.supervisor ORDER BY supervisorshelper.pos) AS json_agg_1\n           FROM (user_supervisors\n             JOIN supervisorshelper ON ((user_supervisors.id = supervisorshelper.list_id)))\n          WHERE ((user_supervisors.user_id = slurm_jobs.sarc_user_id) AND (user_supervisors.valid @> slurm_jobs.submit_time))) AS supervisors,\n    gpurgudb.rgu AS gpu_type_rgu,\n    gpurgudb.drac_rgu AS gpu_type_rgu_drac,\n    ((COALESCE(slurm_jobs.requested_gres_gpu, (0)::bigint))::double precision * gpurgudb.rgu) AS requested_rgu,\n    ((COALESCE(slurm_jobs.requested_gres_gpu, (0)::bigint))::double precision * gpurgudb.drac_rgu) AS requested_rgu_drac,\n    ((COALESCE(slurm_jobs.allocated_gres_gpu, (0)::bigint))::double precision * gpurgudb.rgu) AS allocated_rgu,\n    ((COALESCE(slurm_jobs.allocated_gres_gpu, (0)::bigint))::double precision * gpurgudb.drac_rgu) AS allocated_rgu_drac,\n    (slurm_jobs.elapsed_time * (slurm_jobs.requested_cpu)::double precision) AS requested_cpu_cost,\n    ((((1)::double precision - jobstatisticdb_1.mean) * slurm_jobs.elapsed_time) * (slurm_jobs.requested_cpu)::double precision) AS requested_cpu_waste,\n    (slurm_jobs.elapsed_time * (slurm_jobs.allocated_cpu)::double precision) AS allocated_cpu_cost,\n    ((((1)::double precision - jobstatisticdb_1.mean) * slurm_jobs.elapsed_time) * (slurm_jobs.allocated_cpu)::double precision) AS allocated_cpu_waste,\n    (slurm_jobs.elapsed_time * ((slurm_jobs.allocated_cpu - slurm_jobs.requested_cpu))::double precision) AS cpu_overbilling_cost,\n    ((slurm_jobs.elapsed_time * (slurm_jobs.requested_gres_gpu)::double precision) * gpurgudb.drac_rgu) AS requested_gpu_cost,\n    (((((1)::double precision - jobstatisticdb_2.mean) * slurm_jobs.elapsed_time) * (slurm_jobs.requested_gres_gpu)::double precision) * gpurgudb.drac_rgu) AS requested_gpu_waste,\n    ((slurm_jobs.elapsed_time * (slurm_jobs.allocated_gres_gpu)::double precision) * gpurgudb.drac_rgu) AS allocated_gpu_cost,\n    (((((1)::double precision - jobstatisticdb_2.mean) * slurm_jobs.elapsed_time) * (slurm_jobs.allocated_gres_gpu)::double precision) * gpurgudb.drac_rgu) AS allocated_gpu_waste,\n    ((slurm_jobs.elapsed_time * ((slurm_jobs.allocated_gres_gpu - slurm_jobs.requested_gres_gpu))::double precision) * gpurgudb.drac_rgu) AS gpu_overbilling_cost\n   FROM (((((slurm_jobs\n     LEFT JOIN users ON ((slurm_jobs.sarc_user_id = users.id)))\n     JOIN clusters ON ((slurm_jobs.cluster_id = clusters.id)))\n     LEFT JOIN gpurgudb ON (((gpurgudb.name)::text = (slurm_jobs.harmonized_gpu_type)::text)))\n     LEFT JOIN jobstatisticdb jobstatisticdb_2 ON (((jobstatisticdb_2.job_id = slurm_jobs.id) AND ((jobstatisticdb_2.name)::text = 'gpu_sm_occupancy'::text))))\n     LEFT JOIN jobstatisticdb jobstatisticdb_1 ON (((jobstatisticdb_1.job_id = slurm_jobs.id) AND ((jobstatisticdb_1.name)::text = 'cpu_utilization'::text))))",
-    )
-    op.replace_entity(public_job_series_view)
-    # ### end Alembic commands ###
