--- alembic/versions/dbf2cc8b0356_.py
@@ -1,93 +1,0 @@
-"""Covering indexes for dashboard
-
-Revision ID: dbf2cc8b0356
-Revises: da6dcad6b9f2
-Create Date: 2026-07-14 15:04:38.427515+00:00
-
-"""
-
-from typing import Sequence, Union
-
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision: str = "dbf2cc8b0356"
-down_revision: Union[str, Sequence[str], None] = "da6dcad6b9f2"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f("ix_jobstatisticdb_name"), table_name="jobstatisticdb")
-    op.create_index(
-        "ix_jobstatisticdb_name_job_covering",
-        "jobstatisticdb",
-        ["name", "job_id"],
-        unique=False,
-        postgresql_include=["mean", "max"],
-    )
-    op.drop_index(op.f("ix_slurm_jobs_submit_time"), table_name="slurm_jobs")
-    op.create_index(
-        "ix_slurm_jobs_gpu_submit",
-        "slurm_jobs",
-        ["submit_time"],
-        unique=False,
-        postgresql_include=[
-            "id",
-            "harmonized_gpu_type",
-            "allocated_gres_gpu",
-            "elapsed_time",
-            "cluster_id",
-            "cluster_user",
-            "sarc_user_id",
-        ],
-        postgresql_where=sa.text("allocated_gpu_type IS NOT NULL"),
-    )
-    op.create_index(
-        "ix_slurm_jobs_submit_id",
-        "slurm_jobs",
-        ["submit_time"],
-        unique=False,
-        postgresql_include=["id", "sarc_user_id"],
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(
-        "ix_slurm_jobs_submit_id",
-        table_name="slurm_jobs",
-        postgresql_include=["id", "sarc_user_id"],
-    )
-    op.drop_index(
-        "ix_slurm_jobs_gpu_submit",
-        table_name="slurm_jobs",
-        postgresql_include=[
-            "id",
-            "harmonized_gpu_type",
-            "allocated_gres_gpu",
-            "elapsed_time",
-            "cluster_id",
-            "cluster_user",
-            "sarc_user_id",
-        ],
-        postgresql_where=sa.text("allocated_gpu_type IS NOT NULL"),
-    )
-    op.create_index(
-        op.f("ix_slurm_jobs_submit_time"), "slurm_jobs", ["submit_time"], unique=False
-    )
-    op.drop_index(
-        "ix_jobstatisticdb_name_job_covering",
-        table_name="jobstatisticdb",
-        postgresql_include=["mean", "max"],
-    )
-    op.create_index(
-        op.f("ix_jobstatisticdb_name"), "jobstatisticdb", ["name"], unique=False
-    )
-    # ### end Alembic commands ###
